Atualização de codigo para versão final (apenas voz) e atualização requirements
